refactor(adapters): refactor Hyperswap onto reusable UniswapV3 factory/quoter adapter
- Replaced custom Hyperswap adapter implementation with a subclass of a new generic UniswapV3FactoryQuoterAdapter.
- Added UniswapV3FactoryQuoterAdapter to encapsulate common logic for UniswapV3 style dex quoting via factory and quoter contracts.
- Simplified Hyperswap adapter to configuration only, enabling reuse and direct router integrations for future dexes like Hybra, Upheaval, HX.
- Updated docs and cache handling accordingly.
